# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.HTTPBody:
    properties:
      message:
        type: string
    type: object
  controller.HTTPResponse:
    properties:
      body:
        $ref: '#/definitions/controller.HTTPBody'
      error:
        type: string
      statusCode:
        type: integer
    type: object
  entities.User:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:9191
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Echo Swagger Example API
  version: "1.0"
paths:
  /book/{id}:
    get:
      consumes:
      - application/json
      description: Get book by id
      parameters:
      - description: Api Key
        in: header
        name: api-key
        required: true
        type: string
      - description: Book ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.HTTPResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.HTTPResponse'
      summary: Get book by id
  /book/random:
    get:
      consumes:
      - application/json
      description: Get random books Url
      parameters:
      - description: Api Key
        in: header
        name: api-key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.HTTPResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.HTTPResponse'
      summary: Get books Url
  /user/hash:
    post:
      consumes:
      - application/json
      description: Post user Url
      parameters:
      - description: login
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/entities.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.HTTPResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.HTTPResponse'
      summary: Post user Url
schemes:
- http
swagger: "2.0"